pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["crash1522 <crash1522@naver.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.13"
celery = "^5.4.0"
django = "^5.1.5"
django-ninja = "^1.3.0"
boto3 = "^1.36.10"
langchain = "^0.3.17"
python-dotenv = "^1.0.1"
langchain-openai = "^0.3.3"
langchain-anthropic = "^0.3.5"
langchain-core = "^0.3.33"
django-cors-headers = "^4.3.1"
pydantic = "^2.10.6"
psycopg2-binary = "^2.9.9"
pyyaml = "^6.0.2"
redis = "^5.2.1"

[tool.poetry.group.dev.dependencies]
pytest = "^7.0"
black = "^22.0"
flake8 = "^4.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
